+++ src/api.sc2-directstrike/Controllers/ReplayPlayerController.cs
@@ -20,7 +20,7 @@ public class ReplayPlayerController : ControllerBase
     }
 
     [HttpGet("{id}")]
-    public async Task<ReplayPlayer?> GetById(string pkt, int id)
+    public async Task<ReplayPlayer?> GetById(string pkt, ulong id)
     {
         if (pkt.Length != 24)
         {
@@ -42,8 +42,8 @@ public class ReplayPlayerController : ControllerBase
 
     [HttpGet]
     public async Task<IEnumerable<ReplayPlayer>> Get(string pkt,
-                                                     [FromQuery(Name = "replayId")] int? replayId = null,
-                                                     [FromQuery(Name = "playerId")] int? playerId = null)
+                                                     [FromQuery(Name = "replayId")] ulong? replayId = null,
+                                                     [FromQuery(Name = "playerId")] ulong? playerId = null)
     {
         if (pkt.Length != 24)
         {
@@ -76,7 +76,7 @@ public class ReplayPlayerController : ControllerBase
             ReplayId = replay.Id,
             PlayerId = player.Id,
         };
-        uint id = await dbContext.WriteToDb(pkt, NAME, replayPlayer);
+        ulong id = await dbContext.WriteToDb(pkt, NAME, replayPlayer);
         
         return replayPlayer with { Id = id };
     }
